Show formatted weather today with weather icon.

+++ app/src/main/java/com/dimowner/simpleweather/data/remote/model/Wind.kt
@@ -23,7 +23,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
 		@SerializedName("speed")
-		val speed: Int,
+		val speed: Float,
 		@SerializedName("deg")
-		val deg: Int
+		val deg: Float
 )
